refactor: update role display name logic and introduce country-specific role mappings
